Removed the separate response writer class and used a lambda expression instead.
