+++ src/Microsoft.Health.Fhir.Api/Registration/FhirServerApplicationBuilderExtensions.cs
@@ -3,9 +3,14 @@
 // Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // -------------------------------------------------------------------------------------------------
 
+using System;
+using System.Collections.Generic;
+using System.Net.Mime;
 using EnsureThat;
 using Microsoft.AspNetCore.Diagnostics.HealthChecks;
-using Microsoft.Health.CosmosDb.Features.Health;
+using Microsoft.AspNetCore.Http;
+using Microsoft.Extensions.Diagnostics.HealthChecks;
+using Newtonsoft.Json;
 
 namespace Microsoft.AspNetCore.Builder
 {
@@ -22,7 +27,30 @@ namespace Microsoft.AspNetCore.Builder
 
             app.UseHealthChecks("/health/check", new HealthCheckOptions
             {
-                ResponseWriter = HealthCheckResponseWriter.WriteJson,
+                ResponseWriter = async (httpContext, healthReport) =>
+                {
+                    var allEntities = new List<object>();
+                    foreach (var reportEntry in healthReport.Entries)
+                    {
+                        var individualEntity = new
+                        {
+                            name = reportEntry.Key,
+                            status = Enum.GetName(typeof(HealthStatus), reportEntry.Value.Status),
+                        };
+
+                        allEntities.Add(individualEntity);
+                    }
+
+                    var response = JsonConvert.SerializeObject(
+                        new
+                        {
+                            overallStatus = healthReport.Status.ToString(),
+                            details = allEntities,
+                        });
+
+                    httpContext.Response.ContentType = MediaTypeNames.Application.Json;
+                    await httpContext.Response.WriteAsync(response);
+                },
             });
 
             app.UseStaticFiles();
